api/requirements.txt
django-cors-headers==4.3.0
celery==5.3.0
Django==4.2.1
djangorestframework==3.14.0
python-dotenv==1.0.0
redis==4.5.5
requests==2.31.0
xmltodict==0.13.0
gevent>=24.2.1
gunicorn>=22.0.0
psycopg2-binary>=2.9.6
pycryptodome==3.18.0
PyJWT==2.7.0
django-redis==5.4.0
bcrypt==5.0.0
celery-redbeat==2.3.3